#2988 refactor workflowstatus to state file
